+++ src/main/java/hudson/plugins/performance/actions/PerformanceProjectAction.java
@@ -16,7 +16,7 @@ import hudson.plugins.performance.details.TestSuiteReportDetail;
 import hudson.plugins.performance.details.TrendReportDetail;
 import hudson.plugins.performance.reports.PerformanceReport;
 import hudson.plugins.performance.reports.UriReport;
-import hudson.plugins.performance.reports.throughput.ThroughputReport;
+import hudson.plugins.performance.reports.ThroughputReport;
 import hudson.util.ChartUtil;
 import hudson.util.ChartUtil.NumberOnlyBuildLabel;
 import hudson.util.ColorPalette;